Fix divide

+++ src/divide.cc
@@ -3,5 +3,5 @@
 
 double divide(int x, int y)
 {
-    return x / y;
+    return (double)x / y;
 }
